Reorganize project Structure (non-functional commit)
This commit establishes the temoa package within the temoa folder for purposes of segregating the temoa model code from the testing code and establishing a top-level package beneath the project top level.  All project source code is moved to the temoa package.  Also added output_files folder to catch output and tests folder for testing
