+++ .gitignore
@@ -4,3 +4,6 @@ data_files/debug_logs/lp_files/*
 data_files/untracked_data/
 .DS_Store
 *.ipynb_checkpoints
+
+# ignore the .idea folder from pycharm -- a pycharm specific config folder
+.idea/*
