+++ output_files/.gitignore
@@ -1,0 +1,2 @@
+# ignore everything in the outputs
+*
