+++ temoa/temoa_model/config_sample
