+++ temoa/temoa_model/get_comm_tech.py
