+++ temoa/temoa_model/temoa_myopic.py
