+++ temoa/temoa_model/temoa_rules.py
